requirements.txt
aiosqlite~=0.19.0
aiogram~=2.25.1
python-dotenv~=1.0.0